Added AdditionalActionProvider for CMake
- Extracted import action to BazelImportCurrentProjectAction
- And project detection logic into static method of BazelImportCurrentProjectAction
