+++ base/src/com/google/idea/blaze/base/wizard2/BazelNotificationProvider.java
@@ -1,20 +1,22 @@
+/*
+ * Copyright 2024 The Bazel Authors. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.google.idea.blaze.base.wizard2;
 
-import com.google.idea.blaze.base.bazel.BazelWorkspaceRootProvider;
 import com.google.idea.blaze.base.lang.buildfile.language.BuildFileType;
-import com.google.idea.blaze.base.model.primitives.WorkspaceRoot;
-import com.google.idea.blaze.base.settings.Blaze;
-import com.google.idea.blaze.base.settings.BlazeImportSettings.ProjectType;
-import com.google.idea.blaze.base.settings.BuildSystemName;
-import com.google.idea.blaze.base.sync.workspace.WorkspacePathResolver;
-import com.google.idea.blaze.base.sync.workspace.WorkspacePathResolverImpl;
-import com.intellij.ide.DataManager;
-import com.intellij.ide.util.projectWizard.WizardContext;
-import com.intellij.openapi.actionSystem.ActionPlaces;
-import com.intellij.openapi.actionSystem.AnAction;
-import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.actionSystem.DataContext;
-import com.intellij.openapi.actionSystem.ex.ActionUtil;
 import com.intellij.openapi.fileEditor.FileEditor;
 import com.intellij.openapi.project.DumbAware;
 import com.intellij.openapi.project.Project;
@@ -22,7 +24,6 @@ import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.ui.EditorNotificationPanel;
 import com.intellij.ui.EditorNotificationPanel.Status;
 import com.intellij.ui.EditorNotificationProvider;
-import java.io.File;
 import java.util.function.Function;
 import javax.swing.JComponent;
 import org.jetbrains.annotations.NotNull;
@@ -30,60 +31,13 @@ import org.jetbrains.annotations.Nullable;
 
 public class BazelNotificationProvider implements EditorNotificationProvider, DumbAware {
 
-  private static class ImportAction extends AnAction {
-
-    final File workspaceRootFile;
-
-    ImportAction(File workspaceRootFile) {
-      this.workspaceRootFile = workspaceRootFile;
-    }
-
-    @Override
-    public void actionPerformed(@NotNull AnActionEvent anActionEvent) {
-      BlazeNewProjectWizard wizard = new BlazeNewProjectWizard() {
-        protected void init() {
-          BazelWorkspaceRootProvider.INSTANCE.isWorkspaceRoot(workspaceRootFile);
-          WorkspaceRoot root = new WorkspaceRoot(workspaceRootFile);
-
-          WorkspacePathResolver pathResolver = new WorkspacePathResolverImpl(root);
-          builder.builder().setWorkspaceData(
-              WorkspaceTypeData.builder()
-                  .setWorkspaceName(workspaceRootFile.getName())
-                  .setWorkspaceRoot(root)
-                  .setCanonicalProjectDataLocation(workspaceRootFile)
-                  .setFileBrowserRoot(workspaceRootFile)
-                  .setWorkspacePathResolver(pathResolver)
-                  .setBuildSystem(BuildSystemName.Bazel)
-                  .build()
-          );
-
-          super.init();
-        }
-
-        @Override
-        protected ProjectImportWizardStep[] getSteps(WizardContext context) {
-          return new ProjectImportWizardStep[]{
-              new BlazeSelectProjectViewImportWizardStep(context),
-              new BlazeEditProjectViewImportWizardStep(context)
-          };
-        }
-      };
-
-      if (!wizard.showAndGet()) {
-        return;
-      }
-      BlazeProjectCreator projectCreator = new BlazeProjectCreator(wizard.builder);
-      BlazeImportProjectAction.createFromWizard(projectCreator, wizard.context);
-    }
-  }
-
   @Override
   public @Nullable Function<? super FileEditor, ? extends JComponent> collectNotificationData(
       @NotNull Project project, @NotNull VirtualFile file) {
     if (file.getFileType() != BuildFileType.INSTANCE) {
       return null;
     }
-    if (Blaze.getProjectType(project) != ProjectType.UNKNOWN) {
+    if (!BazelImportCurrentProjectAction.projectCouldBeImported(project)) {
       return null;
     }
 
@@ -92,23 +46,12 @@ public class BazelNotificationProvider implements EditorNotificationProvider, Du
       return null;
     }
 
-    File rootFile = new File(root);
-    if (!BazelWorkspaceRootProvider.INSTANCE.isWorkspaceRoot(new File(root))) {
-      return null;
-    }
-
-    ImportAction action = new ImportAction(rootFile);
-
     return fileEditor -> {
       EditorNotificationPanel panel = new EditorNotificationPanel(fileEditor, Status.Warning);
+      Runnable importAction = BazelImportCurrentProjectAction.createAction(panel, root);
+
       panel.setText("Project is not configured");
-      panel.createActionLabel("Import Bazel project", () -> {
-        DataContext dataContext = DataManager.getInstance().getDataContext(panel);
-        AnActionEvent event = AnActionEvent.createFromAnAction(action, null, ActionPlaces.UNKNOWN, dataContext);
-        if (ActionUtil.lastUpdateAndCheckDumb(action, event, true)) {
-          ActionUtil.performActionDumbAwareWithCallbacks(action, event);
-        }
-      });
+      panel.createActionLabel("Import Bazel project", importAction);
 
       return panel;
     };
