+++ cpp/src/META-INF/blaze-cpp.xml
@@ -68,4 +68,8 @@
   <extensions defaultExtensionNs="com.google.idea.blaze.qsync">
     <BlazeProjectListenerProvider implementation="com.google.idea.blaze.cpp.qsync.CcProjectModelUpdater$Provider" />
   </extensions>
+
+  <extensions defaultExtensionNs="com.intellij.cmake">
+    <notificationActionProvider implementation="com.google.idea.blaze.cpp.BazelCMakeNotificationProvider"/>
+  </extensions>
 </idea-plugin>
